+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add docs for environment variables
 - Improve data provider output
 - Add .env variable `DEFAULT_PATH`
+- Improve duplicated function names output
 
 ## [0.9.0](https://github.com/TypedDevs/bashunit/compare/0.8.0...0.9.0) - 2023-10-15